# Add a maximum allocation size when decoding