--- a/README.md
+++ b/README.md
@@ -80,6 +80,39 @@ fn main() -> Result<(), Error> {
 }
 ```
 
+### Ill-formed data
+
+In order to ease decoding, the Binary Encoding specification of Avro data
+requires some fields to have their length encoded alongside the data.
+
+If encoded data passed to a `Reader` has been ill-formed, it can happen that
+the bytes meant to contain the length of data are bogus and could result
+in extravagant memory allocation.
+
+To shield users from ill-formed data, `avro-rs` sets a limit (default: 512MB)
+to any allocation it will perform when decoding data.
+
+If you expect some of your data fields to be larger than this limit, be sure
+to make use of the `max_allocation_bytes` function before reading **any** data
+(we leverage Rust's [`std::sync::Once`](https://doc.rust-lang.org/std/sync/struct.Once.html) mechanism to initialize this value, if
+any call to decode is made before a call to `max_allocation_bytes`, the limit
+will be 512MB throughout the lifetime of the program).
+
+
+```rust
+extern crate avro_rs;
+
+use avro_rs::max_allocation_bytes;
+
+
+fn main() {
+    max_allocation_bytes(2 * 1024 * 1024 * 1024);  // 2GB
+
+    // ... happily decode large data
+}
+
+```
+
 ## License
 This project is licensed under [MIT License](https://github.com/flavray/avro-rs/blob/master/LICENSE).
 Please note that this is not an official project maintained by [Apache Avro](https://avro.apache.org/).