--- a/src/de.rs
+++ b/src/de.rs
@@ -316,11 +316,9 @@ impl<'de> de::MapAccess<'de> for MapDeserializer<'de> {
         K: DeserializeSeed<'de>,
     {
         match self.input_keys.next() {
-            Some(ref key) => {
-                seed.deserialize(StringDeserializer {
-                    input: (*key).clone(),
-                }).map(Some)
-            },
+            Some(ref key) => seed.deserialize(StringDeserializer {
+                input: (*key).clone(),
+            }).map(Some),
             None => Ok(None),
         }
     }