--- a/src/reader.rs
+++ b/src/reader.rs
@@ -57,8 +57,7 @@ impl<R: Read> Block<R> {
 
         if let Value::Map(meta) = decode(&meta_schema, &mut self.reader)? {
             // TODO: surface original parse schema errors instead of coalescing them here
-            let schema = meta
-                .get("avro.schema")
+            let schema = meta.get("avro.schema")
                 .and_then(|bytes| {
                     if let Value::Bytes(ref bytes) = *bytes {
                         from_slice(bytes.as_ref()).ok()
@@ -73,8 +72,7 @@ impl<R: Read> Block<R> {
                 return Err(ParseSchemaError::new("unable to parse schema").into())
             }
 
-            if let Some(codec) = meta
-                .get("avro.codec")
+            if let Some(codec) = meta.get("avro.codec")
                 .and_then(|codec| {
                     if let Value::Bytes(ref bytes) = *codec {
                         from_utf8(bytes.as_ref()).ok()