--- a/src/schema.rs
+++ b/src/schema.rs
@@ -149,8 +149,7 @@ impl Name {
         if self.name.contains('.') {
             self.name.clone()
         } else {
-            let namespace = self
-                .namespace
+            let namespace = self.namespace
                 .as_ref()
                 .map(|s| s.as_ref())
                 .or(default_namespace);