--- a/src/util.rs
+++ b/src/util.rs
@@ -1,7 +1,29 @@
 use std::io::Read;
+use std::sync::{Once, ONCE_INIT};
 
 use failure::Error;
 use serde_json::{Map, Value};
+
+/// Maximum number of bytes that can be allocated when decoding
+/// Avro-encoded values. This is a protection against ill-formed
+/// data, whose length field might be interpreted as enourmous.
+/// See max_allocation_bytes to change this limit.
+pub static mut MAX_ALLOCATION_BYTES: usize = 512 * 1024 * 1024;
+static MAX_ALLOCATION_BYTES_ONCE: Once = ONCE_INIT;
+
+/// Describes errors happened trying to allocate too many bytes
+#[derive(Fail, Debug)]
+#[fail(display = "Allocation error: {}", _0)]
+pub struct AllocationError(String);
+
+impl AllocationError {
+    pub fn new<S>(msg: S) -> AllocationError
+    where
+        S: Into<String>,
+    {
+        AllocationError(msg.into())
+    }
+}
 
 /// Describes errors happened while decoding Avro data.
 #[derive(Fail, Debug)]
@@ -85,8 +107,12 @@ fn decode_variable<R: Read>(reader: &mut R) -> Result<u64, Error> {
 
     let mut j = 0;
     loop {
+        if j > 9 {
+            // if j * 7 > 64
+            return Err(DecodeError::new("Overflow when decoding integer value").into())
+        }
         reader.read_exact(&mut buf[..])?;
-        i |= (u64::from(buf[0] & 0x7F)) << (j * 7); // TODO overflow
+        i |= (u64::from(buf[0] & 0x7F)) << (j * 7);
         if (buf[0] >> 7) == 0 {
             break
         } else {
@@ -95,6 +121,35 @@ fn decode_variable<R: Read>(reader: &mut R) -> Result<u64, Error> {
     }
 
     Ok(i)
+}
+
+/// Set a new maximum number of bytes that can be allocated when decoding data.
+/// Once called, the limit cannot be changed.
+///
+/// **NOTE** This function must be called before decoding **any** data. The
+/// library leverages [`std::sync::Once`](https://doc.rust-lang.org/std/sync/struct.Once.html)
+/// to set the limit either when calling this method, or when decoding for
+/// the first time.
+pub fn max_allocation_bytes(num_bytes: usize) -> usize {
+    unsafe {
+        MAX_ALLOCATION_BYTES_ONCE.call_once(|| {
+            MAX_ALLOCATION_BYTES = num_bytes;
+        });
+        MAX_ALLOCATION_BYTES
+    }
+}
+
+pub fn safe_len(len: usize) -> Result<usize, Error> {
+    let max_bytes = max_allocation_bytes(512 * 1024 * 1024);
+
+    if len <= max_bytes {
+        Ok(len)
+    } else {
+        Err(AllocationError::new(format!(
+            "Unable to allocate {} bytes (Maximum allowed: {})",
+            len, max_bytes
+        )).into())
+    }
 }
 
 #[cfg(test)]
@@ -108,5 +163,17 @@ mod tests {
         zig_i32(42i32, &mut a);
         zig_i64(42i64, &mut b);
         assert_eq!(a, b);
+    }
+
+    #[test]
+    fn test_overflow() {
+        let causes_left_shift_overflow: &[u8] = &[0xe1, 0xe1, 0xe1, 0xe1, 0xe1];
+        assert!(decode_variable(&mut &causes_left_shift_overflow[..]).is_err());
+    }
+
+    #[test]
+    fn test_safe_len() {
+        assert_eq!(42usize, safe_len(42usize).unwrap());
+        assert!(safe_len(1024 * 1024 * 1024).is_err());
     }
 }